Laid out the API to break it up
